# ARROW-1199: [C++] Implement mutable POD struct for Array data