… cpp/CMakeLists.txt
… cpp/src/arrow/CMakeLists.txt
… cpp/src/arrow/array-test.cc
… cpp/src/arrow/builder.cc
… cpp/src/arrow/compare.cc
… cpp/src/arrow/ipc/feather-test.cc
… cpp/src/arrow/ipc/feather.cc
… cpp/src/arrow/ipc/json-internal.cc
… cpp/src/arrow/ipc/metadata.h
… cpp/src/arrow/ipc/reader.cc
… cpp/src/arrow/ipc/test-common.h
… cpp/src/arrow/ipc/writer.cc
… cpp/src/arrow/loader.cc
… cpp/src/arrow/loader.h
… cpp/src/arrow/pretty_print.cc
… cpp/src/arrow/pretty_print.h
… cpp/src/arrow/python/pandas_convert.cc
… cpp/src/arrow/table.cc
… cpp/src/arrow/table.h
… python/doc/source/development.rst
… python/pyarrow/array.pxi
… python/pyarrow/includes/libarrow.pxd
… python/pyarrow/tests/test_convert_pandas.py