I have been thinking about this one for a little while and I think this is a move in a good direction, with a few open questions discussed below. The `arrow::Array` classes provide user-focused data accessor types. They are not especially convenient for generic data loading and transport, and we have accumulated quite a bit of code for constructing the various array types. They are also not intended for creating functions that mutate memory. For example, suppose we wanted to write a function `Add` that takes two arrays and adds them together into output memory.

One thought is that we could write this as:

```
void Add(FunctionContext* ctx, const Array& left, const Array& right, std::shared_ptr<Array>* out);
```

But this raises a bunch of questions, like what if functions may write into pre-allocated or even resizable memory? This is a little bit out of scope for this change, but trying to motivate reasons why this may be useful.

This introduces an `ArrayData` object that is now part of the base `arrow::Array` class. I provide simplified constructors, both const-ref and rvalue variants (to reduce copying in internal functions).

Performance in microbenchmarks where we construct a lot of arrays is up to 20% worse, but there's more refactorings we can do after this that may reduce overhead in hot paths. 

Before:

```
$ ./release/ipc-read-write-benchmark
Run on (8 X 3661.55 MHz CPU s)
2017-07-10 01:11:18
Benchmark                                                  Time           CPU Iterations
----------------------------------------------------------------------------------------
BM_WriteRecordBatch/1/min_time:1.000/real_time         46541 ns      46541 ns      30274   20.9827GB/s
BM_WriteRecordBatch/4/min_time:1.000/real_time         51885 ns      51886 ns      27351   18.8216GB/s
BM_WriteRecordBatch/16/min_time:1.000/real_time        44582 ns      44582 ns      30753    21.905GB/s
BM_WriteRecordBatch/64/min_time:1.000/real_time        49051 ns      49050 ns      28929   19.9093GB/s
BM_WriteRecordBatch/256/min_time:1.000/real_time       74605 ns      74605 ns      17843   13.0898GB/s
BM_WriteRecordBatch/1024/min_time:1.000/real_time     197055 ns     197054 ns       7031   4.95579GB/s
BM_WriteRecordBatch/4k/min_time:1.000/real_time       660070 ns     660065 ns       2172   1.47948GB/s
BM_WriteRecordBatch/8k/min_time:1.000/real_time      1744683 ns    1744695 ns        757    573.17MB/s
BM_ReadRecordBatch/1/min_time:1.000/real_time           1710 ns       1710 ns     851470   570.991GB/s
BM_ReadRecordBatch/4/min_time:1.000/real_time           3797 ns       3796 ns     365612   257.225GB/s
BM_ReadRecordBatch/16/min_time:1.000/real_time         10749 ns      10749 ns     122955   90.8533GB/s
BM_ReadRecordBatch/64/min_time:1.000/real_time         36022 ns      36023 ns      39487   27.1099GB/s
BM_ReadRecordBatch/256/min_time:1.000/real_time       222198 ns     222189 ns       6110   4.39502GB/s
BM_ReadRecordBatch/1024/min_time:1.000/real_time      935182 ns     935154 ns       1563   1069.31MB/s
BM_ReadRecordBatch/4k/min_time:1.000/real_time       3716467 ns    3716366 ns        372   269.073MB/s
BM_ReadRecordBatch/8k/min_time:1.000/real_time       7425148 ns    7424932 ns        182   134.677MB/s
```

After

```
$ ./release/ipc-read-write-benchmark
Run on (8 X 3699.06 MHz CPU s)
2017-07-10 01:17:23
Benchmark                                                  Time           CPU Iterations
----------------------------------------------------------------------------------------
BM_WriteRecordBatch/1/min_time:1.000/real_time         49833 ns      49828 ns      30273   19.5968GB/s
BM_WriteRecordBatch/4/min_time:1.000/real_time         47210 ns      47210 ns      27166   20.6853GB/s
BM_WriteRecordBatch/16/min_time:1.000/real_time        42273 ns      42273 ns      33185   23.1015GB/s
BM_WriteRecordBatch/64/min_time:1.000/real_time        47820 ns      47818 ns      27915   20.4215GB/s
BM_WriteRecordBatch/256/min_time:1.000/real_time       73806 ns      73806 ns      17881   13.2314GB/s
BM_WriteRecordBatch/1024/min_time:1.000/real_time     203137 ns     203123 ns       6898    4.8074GB/s
BM_WriteRecordBatch/4k/min_time:1.000/real_time       754205 ns     754201 ns       2191   1.29482GB/s
BM_WriteRecordBatch/8k/min_time:1.000/real_time      1769919 ns    1769922 ns        799   564.998MB/s
BM_ReadRecordBatch/1/min_time:1.000/real_time           1723 ns       1723 ns     820125   566.761GB/s
BM_ReadRecordBatch/4/min_time:1.000/real_time           4023 ns       4023 ns     385334   242.763GB/s
BM_ReadRecordBatch/16/min_time:1.000/real_time         13159 ns      13159 ns     106265   74.2109GB/s
BM_ReadRecordBatch/64/min_time:1.000/real_time         46642 ns      46642 ns      30439   20.9373GB/s
BM_ReadRecordBatch/256/min_time:1.000/real_time       224499 ns     224499 ns       5642   4.34996GB/s
BM_ReadRecordBatch/1024/min_time:1.000/real_time     1183760 ns    1183757 ns       1317   844.766MB/s
BM_ReadRecordBatch/4k/min_time:1.000/real_time       4230269 ns    4230265 ns        319   236.392MB/s
BM_ReadRecordBatch/8k/min_time:1.000/real_time       8864950 ns    8864977 ns        150   112.804MB/s
```

So the cost to IPC-reconstruct an 8096-column RecordBatch went from 74ms to 88ms, which isn't totally ideal at a first glance; that's about 180 nanoseconds of additional overhead per column. 

I thought about, and am still thinking about using `std::shared_ptr<ArrayData>` in the array base, and I still think this might not be a bad idea that we could experiment more with. As one substantive change, we might consider using `std::vector<std::shared_ptr<ArrayData>>` inside RecordBatch instead of fully-boxed arrays. We could then "box" the ArrayData on the fly to the right user accessor type.