@wesm 2017-07-10
I refactored to use the shared_ptr approach, which has a lot less copying, and will be better for nested data. Also changed internals of RecordBatch to be based on this simpler data structure. Performance is about the same it seems:

```
$ ./release/ipc-read-write-benchmark 
Run on (8 X 3670.51 MHz CPU s)
2017-07-10 04:12:36
Benchmark                                                  Time           CPU Iterations
----------------------------------------------------------------------------------------
BM_WriteRecordBatch/1/min_time:1.000/real_time         46326 ns      46318 ns      29789   21.0802GB/s
BM_WriteRecordBatch/4/min_time:1.000/real_time         47972 ns      47962 ns      28714   20.3569GB/s
BM_WriteRecordBatch/16/min_time:1.000/real_time        44827 ns      44815 ns      31817    21.785GB/s
BM_WriteRecordBatch/64/min_time:1.000/real_time        58261 ns      58248 ns      25366   16.7619GB/s
BM_WriteRecordBatch/256/min_time:1.000/real_time      116146 ns     116127 ns      12305   8.40806GB/s
BM_WriteRecordBatch/1024/min_time:1.000/real_time     382624 ns     382551 ns       4006   2.55228GB/s
BM_WriteRecordBatch/4k/min_time:1.000/real_time      1450150 ns    1449750 ns       1011   689.584MB/s
BM_WriteRecordBatch/8k/min_time:1.000/real_time      2790335 ns    2789340 ns        426    358.38MB/s
BM_ReadRecordBatch/1/min_time:1.000/real_time           1727 ns       1726 ns     801162   565.495GB/s
BM_ReadRecordBatch/4/min_time:1.000/real_time           3625 ns       3624 ns     386808   269.367GB/s
BM_ReadRecordBatch/16/min_time:1.000/real_time         10840 ns      10837 ns     125800   90.0847GB/s
BM_ReadRecordBatch/64/min_time:1.000/real_time         39714 ns      39709 ns      35407   24.5901GB/s
BM_ReadRecordBatch/256/min_time:1.000/real_time       232941 ns     232885 ns       5995   4.19232GB/s
BM_ReadRecordBatch/1024/min_time:1.000/real_time     1019862 ns    1019622 ns       1343   980.524MB/s
BM_ReadRecordBatch/4k/min_time:1.000/real_time       4185589 ns    4184750 ns        333   238.915MB/s
BM_ReadRecordBatch/8k/min_time:1.000/real_time       8630499 ns    8628216 ns        160   115.868MB/s
```

Probably need to make a FlameGraph to improve this any further