@wesm 2017-07-10
The Parquet build is failing due to some minor API renaming for consistency (data -> values). If you could also review the names for things hopefully we can avoid too many future API breakages like this