@wesm 2017-07-10
Thanks for the review. Another benefit of this work is that the IPC loading machinery becomes substantially simpler. I was able to delete about half of array/loader.h and array/loader.cc and consolidate that code in arrow/ipc which makes more sense