@wesm 2017-07-10
Got to get Python working again (this refactor has actually exposed a number of pre-existing bugs). Then will move onto GLib