@wesm 2017-07-11
OK, I'm going to give Windows a kick locally to see if any issues then merge this, then the Parquet patch, and will circle back if builds are failing