@wesm 2017-07-11
@xhochy I discovered a bug in ArrayBuilder::UnsafeSetNotNull https://github.com/apache/arrow/pull/824/commits/d0a8ee2bc4b3261976e6303c81ed1801bb869dcd; I'm amazed this one evaded us but it suggests our unit testing of the builders is not quite adequate