@wesm 2017-07-11
Merging this so we can try to get a passing Parquet build